% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_groups.R
\name{country_groups}
\alias{country_groups}
\title{Identifies the IMF groups}
\usage{
country_groups(countries)
}
\arguments{
\item{countries}{A vector or a list of countries.}
}
\value{
A list.
}
\description{
Identifies the IMF groups
}
