% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fq_imf_search.R
\name{fq_imf_search}
\alias{fq_imf_search}
\title{Search for the IMF indicators avaiable in Quandl.}
\usage{
fq_imf_search(...)
}
\arguments{
\item{...}{Character strings to match the desired search.}
}
\value{
A \code{tibble} with the indicators' name and the respective \code{Quandl} code.
}
\description{
It covers 44 indicators used in the World Economic Outlook and Fiscal Monitor.
}
\details{
The full list of indicators can be seem here: \url{https://www.quandl.com/data/ODA-IMF-Cross-Country-Macroeconomic-Statistics/documentation/data-organization}.
}
\examples{
# indicators with the 'gdp' letters.
fq_imf_search('gdp')

# the function is case insensitive
fq_imf_search('goVERnment', "ReveNUE")
}
