% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fq_wb.R
\name{fq_wb}
\alias{fq_wb}
\title{Download World Bank Data Using Quandl API}
\usage{
fq_wb(countries, indicators, verbose = TRUE, ...)
}
\arguments{
\item{countries}{A vector or a list of character strings.}

\item{indicators}{A vector or a list of character strings.}

\item{verbose}{Should warning messages be printed? Default is \code{TRUE}.}

\item{...}{Additional arguments to be passed into \code{Quandl} function.}
}
\value{
A tidy \code{tibble} with four columns: \code{date}, \code{country}, \code{indicator} and \code{value}.
}
\description{
A wrapper around \code{Quandl}. Downloads macroeconomic data for several indicators and countries covered by World Bank.
}
\details{
The \code{countries} argument can be passed as an ISO code or as a country name. The only requirement is that the call must be consistent (must contain only ISO codes or country names, but not both).

Sometimes the user may be interesoted in downloading data for certain regions, like Europe, Latin America, Middle East, etc. For that reason, the \code{countries} argument also accepts the following calls:

\itemize{
  \item 'ae'     - Advanced Economics
  \item 'asean5' - Asean Top 5
  \item 'cis'    - Commonwealth and Independent States
  \item 'eda'    - Emerging and Developing Asia
  \item 'ede'    - Emerging and Developing Economies
  \item 'edeuro' - Emerging and Developing Europe
  \item 'euro'   - Euro Area
  \item 'eu'     - European Union
  \item 'g7'     - G7
  \item 'latam'  - Latin America & Caribbean
  \item 'me'     - Middle East
  \item 'oae'    - Other Advanced Economies
  \item 'ssa'    - Sub Saharan Africa
}

For any of those calls the \code{fq_wb()} will download data for all the countries in the requested region. A complete region list can be seen at: \url{https://www.imf.org/external/pubs/ft/weo/2018/01/weodata/groups.htm}.

The \code{...} argument can be used to calibrate the query parameters. It accepts the following calls:

\itemize{
 \item \code{start_date}: \code{'YYYY-MM-DD'}
 \item \code{end_date}: \code{'YYYY-MM-DD'}
 \item \code{collapse}: \code{c("", "daily", "weekly", "monthly", "quarterly", "annual")}
 \item \code{transform}: \code{c("", "diff", "rdiff", "normalize", "cumul", "rdiff_from")}
}

The full list can be seen at: \url{https://docs.quandl.com/docs/r}.

\code{Quandl} has a diverse of World Bank datasets. The following items give a brief description of what you could expect to find when trying to dig further into each one of them:

\itemize{
  \item \href{https://www.quandl.com/data/WWDI-World-Bank-World-Development-Indicators}{World Development Indicators}: Most current and accurate development indicators, compiled from officially-recognized international sources.
  \item \href{https://www.quandl.com/data/WWGI-World-Bank-Worldwide-Governance-Indicators}{Worldwide Governance Indicators}: Data on aggregate and individual governance indicators for six dimensions of governance.
  \item \href{https://www.quandl.com/data/WPSD-World-Bank-Public-Sector-Debt}{Public Sector Debt}: Data jointly developed by the World Bank and the International Monetary Fund, which brings together detailed public sector government debt data.
  \item \href{https://www.quandl.com/data/WPOV-World-Bank-Poverty-Statistics}{Poverty Statistics}: Indicators on poverty headcount ratio, poverty gap, and number of poor at both international and national poverty lines.
  \item \href{https://www.quandl.com/data/WMDG-World-Bank-Millennium-Development-Goals}{Millennium Development Goals}: Data drawn from the World Development Indicators, reorganized according to the goals and targets of the Millennium Development Goals (MDGs).
  \item \href{https://www.quandl.com/data/WJKP-World-Bank-Jobs-for-Knowledge-Platform}{Jobs for Knowledge Platform}: Indicators on labor-related topics.
  \item \href{https://www.quandl.com/data/WIDA-World-Bank-International-Development-Association}{International Development Association}: Data on progress on aggregate outcomes for IDA (International Development Association) countries for selected indicators.
  \item \href{https://www.quandl.com/data/WHNP-World-Bank-Health-Nutrition-and-Population-Statistics}{Health Nutrition and Population Statistics}: Key health, nutrition and population statistics.
  \item \href{https://www.quandl.com/data/WGLF-World-Bank-Global-Findex-Global-Financial-Inclusion-database}{Global Findex (Global Financial Inclusion database)}: Indicators of financial inclusion measures on how people save, borrow, make payments and manage risk.
  \item \href{https://www.quandl.com/data/WGFD-World-Bank-Global-Financial-Development}{Global Financial Development}: Data on financial system characteristics, including measures of size, use, access to, efficiency, and stability of financial institutions and markets.
  \item \href{https://www.quandl.com/data/WGEP-World-Bank-GEP-Economic-Prospects}{GEP Economic Prospects}: Data on the short-, medium, and long-term outlook for the global economy and the implications for developing countries and poverty reduction.
  \item \href{https://www.quandl.com/data/WGEN-World-Bank-Gender-Statistics}{Gender Statistics}: Data describing gender differences in earnings, types of jobs, sectors of work, farmer productivity, and entrepreneurs’ firm sizes and profits.
  \item \href{https://www.quandl.com/data/WGEM-World-Bank-Global-Economic-Monitor}{Global Economic Monitor}: Data on global economic developments, with coverage of high-income, as well as developing countries.
  \item \href{https://www.quandl.com/data/WGEC-World-Bank-Global-Economic-Monitor-GEM-Commodities}{Global Economic Monitor (GEM) Commodities}: Data containing commodity prices and indices from 1960 to present.
  \item \href{https://www.quandl.com/data/WGDF-World-Bank-Global-Development-Finance}{Global Development Finance}: Data on financial system characteristics, including measures of size, use, access to, efficiency, and stability of financial institutions and markets.
  \item \href{https://www.quandl.com/data/WESV-World-Bank-Enterprise-Surveys}{Enterprise Surveys}: Company-level private sector data, covering business topics including finance, corruption, infrastructure, crime, competition, and performance measures.
  \item \href{https://www.quandl.com/data/WEDU-World-Bank-Education-Statistics}{Education Statistics}: Internationally comparable indicators on education access, progression, completion, literacy, teachers, population, and expenditures.
  \item \href{https://www.quandl.com/data/WDBU-World-Bank-Doing-Business}{Doing Business}: Data on business regulations and their enforcement for member countries and selected cities at the subnational and regional level.
  \item \href{https://www.quandl.com/data/WCSC-World-Bank-Corporate-Scorecard}{Corporate Scorecard}: This database is designed to provide a strategic overview of the World Bank Group’s performance toward ending extreme poverty and promoting shared prosperity.
  \item \href{https://www.quandl.com/data/WADI-World-Bank-Africa-Development-Indicators}{Africa Development Indicators}: A collection of development indicators on Africa, including national, regional and global estimates.
}
}
\examples{
# load the package
library(FromQuandl)

# Download the 1 year return of Real effective exchange rate index for all Advanced Economies.
fq_wb('ae', list('PX_REX_REER'), verbose = FALSE, transform = 'rdiff')

# Downlaod the Control of Corruption Estimate in the US
fq_wb('USA', 'CC_EST')

# Download al indicators related to Rule of Law for the G7 countries from 2010 ownwards...
fq_wb(countries  = 'g7',
      indicators = c('RL_EST', 'RL_NO_SRC', 'RL_PER_RNK', 'RL_STD_ERR'),
      start_date = '2010-01-01')
}
