% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fq_yale.R
\name{fq_yale}
\alias{fq_yale}
\title{Yale Department of Economics}
\usage{
fq_yale(indicators, verbose = TRUE, ...)
}
\arguments{
\item{indicators}{A vector or a list of character strings.}

\item{verbose}{Should warning messages be printed? Default is \code{TRUE}.}

\item{...}{Additional arguments to be passed into \code{Quandl} function.}
}
\value{
A tidy \code{tibble} with five columns: \code{date}, \code{code}, \code{indicator}, \code{description} and \code{value}.
}
\description{
Publishes data related to the stock and housing markets; price data for common indexes, interest rates, earnings, dividends, house prices, and CPI.
}
\details{
Data provided for the US economy only.
}
\examples{
# Ten-Year Average of Real Earnings for US Stock Data
fq_yale("SP_10YRE")
}
