% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FromQuandl.R
\docType{package}
\name{FromQuandl}
\alias{FromQuandl}
\title{FromQuandl}
\description{
A package that make it easier to download bundles of data from Quandl.
}
\section{FromQuandl functions}{

\itemize{
  \item \code{\link{fq_imf}}: Downloads macroeconomic data for several indicators and countries covered by the IMF.
  \item \code{\link{fq_wb}}: Downloads macroeconomic data for several indicators and countries covered by World Bank.
  \item \code{\link{fq_yale}}: Downloads macroeconomic data for several indicators covered by the Yale Department of Economics.
}
}

