% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imf_datasets.R
\docType{data}
\name{imf_datasets}
\alias{imf_datasets}
\title{IMF Macroeconomia Indicators}
\format{A data frame with 43 rows and 2 columns:

\describe{
  \item{imf_name}{A Macroeconomic variable}
  \item{imf_code}{Code in Quandl}
  ...
}}
\source{
\url{https://www.quandl.com/data/ODA-IMF-Cross-Country-Macroeconomic-Statistics/documentation/data-organization
}
}
\usage{
imf_datasets
}
\description{
A dataset containing all the macroeconomic variables monitored by the IMF that can be found in Quandl.
}
\keyword{datasets}
