% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imf_search.R
\name{imf_search}
\alias{imf_search}
\title{Search for the IMF indicators avaiable in Quandl.}
\usage{
imf_search(query, print_all = TRUE)
}
\arguments{
\item{query}{A character string.}

\item{print_all}{Logical vector. TRUE if you want to overwrite the tibble printing format.}
}
\value{
A \code{tibble} with the indicators' name and the respective \code{Quandl} code.
}
\description{
It covers 44 indicators used in the World Economic Outlook and Fiscal Monitor.
}
\details{
The full list of indicators can be seem here: \url{https://www.quandl.com/data/ODA-IMF-Cross-Country-Macroeconomic-Statistics/documentation/data-organization}.
}
\examples{
# indicators with the 'gdp' letters.
imf_search('gdp')

# prints only the first 10 results.
# the function is case insensitive
imf_search('goVERnment', print_all = FALSE)
}
