% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_development_from_quandl.R
\name{wb_development_from_quandl}
\alias{wb_development_from_quandl}
\title{Download World Bank Development Data From Quandl}
\usage{
wb_development_from_quandl(countries, indicators, ...)
}
\arguments{
\item{countries}{A vector or a list of character strings.}

\item{indicators}{A vector or a list of character strings.}

\item{...}{Additional parameters to be passed to the Quandl API.}
}
\value{
A tidy \code{tibble}.
}
\description{
A wrapper around \code{Quandl}. Download development data covered by the World Bank with just one line of code.
}
\details{
The \code{countries} argument can be passed as an ISO code or as a country name. The only requirement is that the call must be consistent (must contain only ISO codes or only country names, but not both).

Sometimes the user may be interested in downloading data for certain regions, like Europe, Latin America, Middle East, etc. For that reason, the countries argument also accepts the following calls:

\itemize{
  \item 'ae'     - Advanced Economics
  \item 'asean_5'- Asean Top 5
  \item 'cis'    - Commonwealth and Independent States
  \item 'eda'    - Emerging and Developing Asia
  \item 'ede'    - Emerging and Developing Economies
  \item 'edeuro' - Emerging and Developing Europe
  \item 'euro'   - Euro Area
  \item 'eu'     - European Union
  \item 'g7'     - G7
  \item 'latam'  - Latin America & Caribbean
  \item 'me'     - Middle East
  \item 'oae'    - Other Advanced Economies
  \item 'ssa'    - Sub Saharan Africa
}

For any of those calls the \code{wb_development_from_quandl()} will download data for all the countries in the requested region. A complete region list can be seen at: \url{https://www.imf.org/external/pubs/ft/weo/2018/01/weodata/groups.htm}.

The \code{...} argument can be used to calibrate the query parameters. It accepts the following calls:
\itemize{
 \item \code{start_date}: \code{'YYYY-MM-DD'}
 \item \code{end_date}: \code{'YYYY-MM-DD'}
 \item \code{collapse}: \code{c("", "daily", "weekly", "monthly", "quarterly", "annual")}
 \item \code{transform}: \code{c("", "diff", "rdiff", "normalize", "cumul", "rdiff_from")}
}

The full list can be seen at: \url{https://docs.quandl.com/docs/r}.
}
\examples{
# load package
library(FromQuandl)

# Downlaod the Broad money (\% of GDP)  for all countries in the euro area
wb_development_from_quandl(countries = 'euro', indicators = 'FM_LBL_BMNY_GD_ZS')

# Download the number of children out of school in the primary for countries
# in the Sub Saharan Africa from 1900 up to 2000.
wb_development_from_quandl('ssa', 'SE_PRM_UNER', start_date = '1990-01-01', end_date = '2000-01-01')

# Download and plot the 1 year return of Real effective exchange rate index
# for all Advanced Economies.
wb_development_from_quandl('ae', list('PX_REX_REER'), transform = 'rdiff') \%>\%
  ggplot(aes(x = date, y = value, color = country)) +
  geom_line(show.legend = FALSE) +
  facet_wrap(~ country, scales = 'free_y')

}
