% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_from_quandl.R
\name{wb_from_quandl}
\alias{wb_from_quandl}
\title{Download World Bank data using Quandl API}
\usage{
wb_from_quandl(countries, indicators, ...)
}
\arguments{
\item{countries}{A vector or a list of character strings.}

\item{indicators}{A vector or a list of character strings.}

\item{...}{Additional arguments to be passed into \code{Quandl} function.}
}
\value{
A tidy \code{tibble}.
}
\description{
A wrapper around \code{Quandl}. Download Governance and Poverty data covered by the World Bank with just one line of code.
}
\details{
The \code{countries} argument can be passed as an ISO code or as a country name. The only requirement is that the call must be consistent (must contain only ISO codes or country names, but not both).

Sometimes the user may be interested in downloading data for certain regions, like Europe, Latin America, Middle East, etc. For that reason, the countries argument also accepts the following calls:

\itemize{
  \item advanced_economics
  \item asean_5
  \item commonwealth_independent_states
  \item emerging_and_developing_asia
  \item emerging_and_developing_economies
  \item emerging_and_developing_europe
  \item euro_area
  \item european_union
  \item g7
  \item latin_america_and_caribbean
  \item middle_east
  \item other_advanced_economies
  \item sub_saharan_africa
}

For any of those calls the \code{imf_from_quandl()} will download data for all the countries in the requested region. A complete region list can be seen at: \url{https://www.imf.org/external/pubs/ft/weo/2018/01/weodata/groups.htm}.

The \code{...} argument can be used to calibrate the query parameters. It accepts the following calls:
\itemize{
 \item \code{start_date} and \code{end_date}, to filter time series:
 \item \code{collapse} and \code{transform}, for preprocess
}

The full list can be seen at: \url{https://docs.quandl.com/docs/r}.
}
\examples{
wb_from_quandl(countries = c('CAN', USA'), indicator = 'poverty')

# case insensitive
wb_from_quandl(countries = g7, indicators = 'develOPMENT', start_date = '2016-01-01)
}
