% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_governance_from_quandl.R
\name{wb_governance_from_quandl}
\alias{wb_governance_from_quandl}
\title{Download World Bank Governance Data From Quandl}
\usage{
wb_governance_from_quandl(countries, indicators, ...)
}
\arguments{
\item{countries}{A vector or a list of character strings.}

\item{indicators}{A vector or a list of character strings.}

\item{...}{Additional parameters to be passed to the Quandl API.}
}
\value{
A tidy \code{tibble}.
}
\description{
A wrapper around \code{Quandl}. Download governance data covered by the World Bank with just one line of code.
}
\details{
The \code{countries} argument can be passed as an ISO code or as a country name. The only requirement is that the call must be consistent (must contain only ISO codes or only country names, but not both).

Sometimes the user may be interested in downloading data for certain regions, like Europe, Latin America, Middle East, etc. For that reason, the countries argument also accepts the following calls:

\itemize{
  \item 'ae'     - Advanced Economics
  \item 'asean_5'- Asean Top 5
  \item 'cis'    - Commonwealth and Independent States
  \item 'eda'    - Emerging and Developing Asia
  \item 'ede'    - Emerging and Developing Economies
  \item 'edeuro' - Emerging and Developing Europe
  \item 'euro'   - Euro Area
  \item 'eu'     - European Union
  \item 'g7'     - G7
  \item 'latam'  - Latin America & Caribbean
  \item 'me'     - Middle East
  \item 'oae'    - Other Advanced Economies
  \item 'ssa'    - Sub Saharan Africa
}

For any of those calls the \code{wb_governance_from_quandl()} will download data for all the countries in the requested region. A complete region list can be seen at: \url{https://www.imf.org/external/pubs/ft/weo/2018/01/weodata/groups.htm}.

The \code{...} argument can be used to calibrate the query parameters. It accepts the following calls:
\itemize{
 \item \code{start_date}: \code{'YYYY-MM-DD'}
 \item \code{end_date}: \code{'YYYY-MM-DD'}
 \item \code{collapse}: \code{c("", "daily", "weekly", "monthly", "quarterly", "annual")}
 \item \code{transform}: \code{c("", "diff", "rdiff", "normalize", "cumul", "rdiff_from")}
}

The full list can be seen at: \url{https://docs.quandl.com/docs/r}.
}
\examples{
library(FromQuandl)

# Downlaod the Control of Corruption Estimate in the US
wb_governance_from_quandl('USA', 'CC_EST')

# Download the Government Effectiveness Percentile Rank for all countries in the Middle East
wb_governance_from_quandl('me', 'GE_PER_RNK')

# Download al indicators related to Rule of Law for the G7 countries from 2010 ownwards...
wb_search('law')
wb_governance_from_quandl(countries  = 'g7',
                          indicators = c('RL_EST', 'RL_NO_SRC', 'RL_PER_RNK', 'RL_STD_ERR'),
                          start_date = '2010-01-01')

# Download the Number of Sources of "Voice and Accountability" for all countries
# in Latin America
wb_governance_from_quandl('latam', 'VA_NO_SRC')

}
