% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_poverty_from_quandl.R
\name{wb_poverty_from_quandl}
\alias{wb_poverty_from_quandl}
\title{Download World Bank Poverty Data From Quandl}
\usage{
wb_poverty_from_quandl(countries, indicators, ...)
}
\arguments{
\item{countries}{A vector or a list of character strings.}

\item{indicators}{A vector or a list of character strings.}

\item{...}{Additional parameters to be passed to the Quandl API.}
}
\value{
A tidy \code{tibble}.
}
\description{
A wrapper around \code{Quandl}. Download Poverty data covered by the World Bank with just one line of code.
}
\details{
The \code{countries} argument can be passed as an ISO code or as a country name. The only requirement is that the call must be consistent (must contain only ISO codes or only country names, but not both).

Sometimes the user may be interested in downloading data for certain regions, like Europe, Latin America, Middle East, etc. For that reason, the countries argument also accepts the following calls:

\itemize{
  \item 'ae'     - Advanced Economics
  \item 'asean_5'- Asean Top 5
  \item 'cis'    - Commonwealth and Independent States
  \item 'eda'    - Emerging and Developing Asia
  \item 'ede'    - Emerging and Developing Economies
  \item 'edeuro' - Emerging and Developing Europe
  \item 'euro'   - Euro Area
  \item 'eu'     - European Union
  \item 'g7'     - G7
  \item 'latam'  - Latin America & Caribbean
  \item 'me'     - Middle East
  \item 'oae'    - Other Advanced Economies
  \item 'ssa'    - Sub Saharan Africa
}

For any of those calls the \code{wb_poverty_from_quandl()} will download data for all the countries in the requested region. A complete region list can be seen at: \url{https://www.imf.org/external/pubs/ft/weo/2018/01/weodata/groups.htm}.

The \code{...} argument can be used to calibrate the query parameters. It accepts the following calls:
\itemize{
 \item \code{start_date}: \code{'YYYY-MM-DD'}
 \item \code{end_date}: \code{'YYYY-MM-DD'}
 \item \code{collapse}: \code{c("", "daily", "weekly", "monthly", "quarterly", "annual")}
 \item \code{transform}: \code{c("", "diff", "rdiff", "normalize", "cumul", "rdiff_from")}
}

The full list can be seen at: \url{https://docs.quandl.com/docs/r}.
}
\examples{
library(FromQuandl)

# Plot and model the poor at $2 a day (PPP) in
# emerging and Developing Europe Countries
if (requireNamespace('ggplot2',quietly = TRUE)) {
    wb_poverty_from_quandl(
      countries  = 'edeuro',
      indicators ='SI_POV_NOP2',
      start_date = '2000-01-01',
      end_date   = '2018-01-01'
    ) \%>\%
    ggplot(aes(date, value, color = country)) +
    geom_point(show.legend = FALSE) +
    geom_smooth(show.legend = FALSE) +
    facet_wrap(~ country, scales = 'free_y') +
    theme_classic()
}
}
