% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_search.R
\name{wb_search}
\alias{wb_search}
\title{Search for the World Bank indicators avaiable in Quandl.}
\usage{
wb_search(query, print_all = TRUE)
}
\arguments{
\item{query}{A vector or a list of character strings.}

\item{print_all}{Logical vector. If TRUE the all the results will be printed overwriting the standard \code{tibble} format.}
}
\value{
A \code{tibble} with the indicators' name and the respective \code{Quandl} code.
}
\description{
It covers 1309 indicators spread among the Governance, Development, Poverty and Public Sector databases.
}
\details{
The full list of indicators can be seen in the following links:

\itemize{
    \item Governance: \url{https://www.quandl.com/data/WWGI-World-Bank-Worldwide-Governance-Indicators/documentation/data-organization}
    \item Development: \url{https://www.quandl.com/data/WWDI-World-Bank-World-Development-Indicators/documentation/data-organization}
    \item Poverty: \url{https://www.quandl.com/data/WPOV-World-Bank-Poverty-Statistics/documentation/data-organization}
    \item Public Sector: \url{https://www.quandl.com/data/WPSD-World-Bank-Public-Sector-Debt}
}
}
\examples{
# oil search
wb_search(query = 'oil')

# The function is case insensitive
wb_search(query = 'EDuCAtion')
}
