% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fq_imf.R
\name{fq_imf}
\alias{fq_imf}
\title{Download IMF data using Quandl API}
\usage{
fq_imf(countries, indicators, verbose = TRUE, ...)
}
\arguments{
\item{countries}{A vector or a list of character strings.}

\item{indicators}{A vector or a list of character strings.}

\item{verbose}{Should warning messages be printed? Default is \code{TRUE}.}

\item{...}{Additional arguments to be passed into \code{Quandl} function.}
}
\value{
A tidy \code{tibble} with four columns: \code{date}, \code{country}, \code{indicator} and \code{value}.
}
\description{
A wrapper around \code{Quandl}. Downloads macroeconomic data for several indicators and countries covered by the IMF.
}
\details{
The \code{countries} argument can be passed as an ISO code or as a country name. The only requirement is that the call must be consistent (must contain only ISO codes or country names, but not both).

Sometimes the user may be interested in downloading data for certain regions, like Europe, Latin America, Middle East, etc. For that reason, the countries argument also accepts the following calls:

\itemize{
  \item 'ae'     - Advanced Economics
  \item 'asean5' - Asean Top 5
  \item 'cis'    - Commonwealth and Independent States
  \item 'eda'    - Emerging and Developing Asia
  \item 'ede'    - Emerging and Developing Economies
  \item 'edeuro' - Emerging and Developing Europe
  \item 'euro'   - Euro Area
  \item 'eu'     - European Union
  \item 'g7'     - G7
  \item 'latam'  - Latin America & Caribbean
  \item 'me'     - Middle East
  \item 'oae'    - Other Advanced Economies
  \item 'ssa'    - Sub Saharan Africa
}

For any of those calls the \code{fq_imf()} will download data for all the countries in the requested region. A complete region list can be seen at: \url{https://www.imf.org/external/pubs/ft/weo/2018/01/weodata/groups.htm}.

The \code{...} argument can be used to calibrate the query parameters. It accepts the following calls:
\itemize{
 \item \code{start_date}: \code{'YYYY-MM-DD'}
 \item \code{end_date}: \code{'YYYY-MM-DD'}
 \item \code{collapse}: \code{c("", "daily", "weekly", "monthly", "quarterly", "annual")}
 \item \code{transform}: \code{c("", "diff", "rdiff", "normalize", "cumul", "rdiff_from")}
}

The full list can be seen at: \url{https://docs.quandl.com/docs/r}.
}
\examples{
# Download the US Output Gap
fq_imf('United States', 'NGAP_NPGDP')

# The example above is identical to
# fq_imf('USA', 'NGAP_NPGDP')
}
