% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FromQuandl.R
\docType{package}
\name{FromQuandl}
\alias{FromQuandl}
\alias{FromQuandl-package}
\title{FromQuandl: A package that make it easier to download bundles of data from Quandl.}
\description{
FromQuandl: A package that make it easier to download bundles of data from Quandl.
}
\section{FromQuandl functions}{

\itemize{
  \item \code{\link{fq_imf}}: Downloads macroeconomic data for several indicators and countries covered by the IMF.
  \item \code{\link{fq_wb}}: Downloads macroeconomic data for several indicators and countries covered by World Bank.
}
}

